refactor: extract profile_hooks.py from __init__.py
Move profile lifecycle handlers (~65 lines) into profile_hooks.py.
Includes tip of the day, monthly pokemon, AnkiWeb sync setup, and
hook registration for profile load/close.
